requirements.txt
flask
numpy
pandas
gunicorn
flask_cors
sqlite3
tensorflow-cpu